requirements.txt
Django==5.2.7
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
Pillow==11.0.0
sqlparse==0.5.3
asgiref==3.10.0
openpyxl==3.1.5
requests==2.32.5
beautifulsoup4==4.14.2
duckduckgo-search==8.1.1

